+++ superset/assets/src/explore/controlPanels/BigNumber.js
@@ -1,0 +1,29 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  controlPanelSections: [
+    {
+      label: t('Query'),
+      expanded: true,
+      controlSetRows: [
+        ['metric'],
+        ['adhoc_filters'],
+      ],
+    },
+    {
+      label: t('Chart Options'),
+      expanded: true,
+      controlSetRows: [
+        ['compare_lag', 'compare_suffix'],
+        ['y_axis_format', null],
+        ['show_trend_line', 'start_y_axis_at_zero'],
+        ['color_picker', null],
+      ],
+    },
+  ],
+  controlOverrides: {
+    y_axis_format: {
+      label: t('Number format'),
+    },
+  },
+};
