+++ superset/assets/src/explore/controlPanels/Compare.js
@@ -1,0 +1,43 @@
+import { t } from '@superset-ui/translation';
+import { NVD3TimeSeries, annotations } from './sections';
+import { D3_TIME_FORMAT_OPTIONS } from '../controls';
+
+export default {
+  requiresTime: true,
+  controlPanelSections: [
+    NVD3TimeSeries[0],
+    {
+      label: t('Chart Options'),
+      expanded: true,
+      controlSetRows: [
+        ['color_scheme'],
+      ],
+    },
+    {
+      label: t('X Axis'),
+      expanded: true,
+      controlSetRows: [
+        ['x_axis_label', 'bottom_margin'],
+        ['x_ticks_layout', 'x_axis_format'],
+        ['x_axis_showminmax', null],
+      ],
+    },
+    {
+      label: t('Y Axis'),
+      expanded: true,
+      controlSetRows: [
+        ['y_axis_label', 'left_margin'],
+        ['y_axis_showminmax', 'y_log_scale'],
+        ['y_axis_format', 'y_axis_bounds'],
+      ],
+    },
+    NVD3TimeSeries[1],
+    annotations,
+  ],
+  controlOverrides: {
+    x_axis_format: {
+      choices: D3_TIME_FORMAT_OPTIONS,
+      default: 'smart_date',
+    },
+  },
+};
