+++ superset/assets/src/explore/controlPanels/CountryMap.js
@@ -1,0 +1,36 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  controlPanelSections: [
+    {
+      label: t('Query'),
+      expanded: true,
+      controlSetRows: [
+        ['entity'],
+        ['metric'],
+        ['adhoc_filters'],
+      ],
+    },
+    {
+      label: t('Options'),
+      expanded: true,
+      controlSetRows: [
+        ['select_country', 'number_format'],
+        ['linear_color_scheme'],
+      ],
+    },
+  ],
+  controlOverrides: {
+    entity: {
+      label: t('ISO 3166-2 codes of region/province/department'),
+      description: t('It\'s ISO 3166-2 of your region/province/department in your table. (see documentation for list of ISO 3166-2)'),
+    },
+    metric: {
+      label: t('Metric'),
+      description: 'Metric to display bottom title',
+    },
+    linear_color_scheme: {
+      renderTrigger: false,
+    },
+  },
+};
