+++ superset/assets/src/explore/controlPanels/DeckPath.js
@@ -1,0 +1,42 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  requiresTime: true,
+  controlPanelSections: [
+    {
+      label: t('Query'),
+      expanded: true,
+      controlSetRows: [
+        ['line_column', 'line_type'],
+        ['row_limit', 'filter_nulls'],
+        ['adhoc_filters'],
+      ],
+    },
+    {
+      label: t('Map'),
+      expanded: true,
+      controlSetRows: [
+        ['mapbox_style', 'viewport'],
+        ['color_picker', 'line_width'],
+        ['reverse_long_lat', 'autozoom'],
+      ],
+    },
+    {
+      label: t('Advanced'),
+      controlSetRows: [
+        ['js_columns'],
+        ['js_data_mutator'],
+        ['js_tooltip'],
+        ['js_onclick_href'],
+      ],
+    },
+  ],
+  controlOverrides: {
+    line_type: {
+      choices: [
+        ['polyline', 'Polyline'],
+        ['json', 'JSON'],
+      ],
+    },
+  },
+};
