+++ superset/assets/src/explore/controlPanels/DeckScreengrid.js
@@ -1,0 +1,49 @@
+import { t } from '@superset-ui/translation';
+import { nonEmpty } from '../validators';
+import timeGrainSqlaAnimationOverrides from './timeGrainSqlaAnimationOverrides';
+
+export default {
+  requiresTime: true,
+  controlPanelSections: [
+    {
+      label: t('Query'),
+      expanded: true,
+      controlSetRows: [
+        ['spatial', 'size'],
+        ['row_limit', 'filter_nulls'],
+        ['adhoc_filters'],
+      ],
+    },
+    {
+      label: t('Map'),
+      controlSetRows: [
+        ['mapbox_style', 'viewport'],
+        ['autozoom', null],
+      ],
+    },
+    {
+      label: t('Grid'),
+      expanded: true,
+      controlSetRows: [
+        ['grid_size', 'color_picker'],
+      ],
+    },
+    {
+      label: t('Advanced'),
+      controlSetRows: [
+        ['js_columns'],
+        ['js_data_mutator'],
+        ['js_tooltip'],
+        ['js_onclick_href'],
+      ],
+    },
+  ],
+  controlOverrides: {
+    size: {
+      label: t('Weight'),
+      description: t("Metric used as a weight for the grid's coloring"),
+      validators: [nonEmpty],
+    },
+    time_grain_sqla: timeGrainSqlaAnimationOverrides,
+  },
+};
