+++ superset/assets/src/explore/controlPanels/Iframe.js
@@ -1,0 +1,12 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  controlPanelSections: [
+    {
+      label: t('Options'),
+      controlSetRows: [
+        ['url'],
+      ],
+    },
+  ],
+};
