+++ superset/assets/src/explore/controlPanels/Markup.js
@@ -1,0 +1,14 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  controlPanelSections: [
+    {
+      label: t('Code'),
+      expanded: true,
+      controlSetRows: [
+        ['markup_type'],
+        ['code'],
+      ],
+    },
+  ],
+};
