+++ superset/assets/src/explore/controlPanels/Treemap.js
@@ -1,0 +1,30 @@
+import { t } from '@superset-ui/translation';
+
+export default {
+  controlPanelSections: [
+    {
+      label: t('Query'),
+      expanded: true,
+      controlSetRows: [
+        ['metrics'],
+        ['adhoc_filters'],
+        ['groupby'],
+        ['row_limit'],
+      ],
+    },
+    {
+      label: t('Chart Options'),
+      expanded: true,
+      controlSetRows: [
+        ['color_scheme'],
+        ['treemap_ratio'],
+        ['number_format'],
+      ],
+    },
+  ],
+  controlOverrides: {
+    color_scheme: {
+      renderTrigger: false,
+    },
+  },
+};
