+++ superset/assets/src/explore/controlPanels/timeGrainSqlaAnimationOverrides.js
@@ -1,0 +1,8 @@
+export default {
+  default: null,
+  mapStateToProps: state => ({
+    choices: (state.datasource) ?
+      state.datasource.time_grain_sqla.filter(o => o[0] !== null) :
+      null,
+  }),
+};
